Add cpp:custom-print
