+++ cpp/src/DataStorm/TraceUtil.h
@@ -49,100 +49,6 @@ namespace std
 }
 // NOLINTEND(cert-dcl58-cpp)
 
-namespace Ice
-{
-    inline std::ostream& operator<<(std::ostream& os, const Ice::Identity& id)
-    {
-        return os << Ice::identityToString(id);
-    }
-}
-
-namespace DataStormContract
-{
-    inline std::string valueIdToString(std::int64_t valueId)
-    {
-        if (valueId < 0)
-        {
-            return "f" + std::to_string(-valueId);
-        }
-        else
-        {
-            return "k" + std::to_string(valueId);
-        }
-    }
-
-    inline std::ostream& operator<<(std::ostream& os, const ElementInfo& info)
-    {
-        os << valueIdToString(info.id);
-        return os;
-    }
-
-    inline std::ostream& operator<<(std::ostream& os, const ElementData& data)
-    {
-        os << 'e' << data.id;
-        if (data.config && data.config->facet)
-        {
-            os << ':' << *data.config->facet;
-        }
-        return os;
-    }
-
-    inline std::ostream& operator<<(std::ostream& os, const ElementSpec& spec)
-    {
-        for (auto q = spec.elements.begin(); q != spec.elements.end(); ++q)
-        {
-            if (q != spec.elements.begin())
-            {
-                os << ',';
-            }
-            os << *q << ':' << valueIdToString(spec.id) << ":pv" << valueIdToString(spec.peerId);
-        }
-        return os;
-    }
-
-    inline std::ostream& operator<<(std::ostream& os, const ElementDataAck& data)
-    {
-        os << 'e' << data.id << ":pe" << data.peerId;
-        if (data.config && data.config->facet)
-        {
-            os << ':' << *data.config->facet;
-        }
-        os << ":sz" << data.samples.size();
-        return os;
-    }
-
-    inline std::ostream& operator<<(std::ostream& os, const ElementSpecAck& spec)
-    {
-        for (auto q = spec.elements.begin(); q != spec.elements.end(); ++q)
-        {
-            if (q != spec.elements.begin())
-            {
-                os << ',';
-            }
-            os << *q << ':' << valueIdToString(spec.id) << ":pv" << valueIdToString(spec.peerId);
-        }
-        return os;
-    }
-
-    inline std::ostream& operator<<(std::ostream& os, const TopicInfo& info)
-    {
-        os << "[" << info.ids << "]:" << info.name;
-        return os;
-    }
-
-    inline std::ostream& operator<<(std::ostream& os, const TopicSpec& info)
-    {
-        os << info.id << ':' << info.name << ":[" << info.elements << "]";
-        return os;
-    }
-
-    inline std::ostream& operator<<(std::ostream& os, const DataSamples& samples)
-    {
-        os << 'e' << samples.id << ":sz" << samples.samples.size();
-        return os;
-    }
-}
-
 namespace DataStormI
 {
     template<typename T, typename std::enable_if_t<std::is_base_of_v<DataStormI::Element, T>>* = nullptr>
