+++ slice/Ice/Version.ice
@@ -24,6 +24,7 @@
 module Ice
 {
     /// A version structure for the protocol version.
+    ["cpp:custom-print"]
     struct ProtocolVersion
     {
         byte major;
@@ -31,6 +32,7 @@ module Ice
     }
 
     /// A version structure for the encoding version.
+    ["cpp:custom-print"]
     struct EncodingVersion
     {
         byte major;
